# Conflicts:
#	pom.xml
#	sp-consumer-order-service/pom.xml
#	uniapp-consumer-ui/manifest.json
#	uniapp-consumer-ui/pages/food/food.vue
#	uniapp-consumer-ui/pages/user/coupons.vue
#	uniapp-consumer-ui/unpackage/dist/dev/mp-weixin/common/vendor.js
#	uniapp-consumer-ui/unpackage/dist/dev/mp-weixin/pages/index/index.js
#	uniapp-consumer-ui/unpackage/dist/dev/mp-weixin/pages/index/index.wxml
#	uniapp-consumer-ui/unpackage/dist/dev/mp-weixin/project.config.json
#	vue-admin-use-tt/src/views/kitchen/comment/qrCode.vue

+++ sp-common/sp-common-model/src/main/java/com/sp/model/domain/Orders.java
@@ -90,7 +90,7 @@ public class Orders implements Serializable {
     /**
      * 实收金额
      */
-    private BigDecimal amount;
+    private double amount;
 
     /**
      * 备注
